[kernels]: share MoE vector bitcast helpers and drop submodule underscores
Hoist the pure vector bitcast repack helpers into kernels/moe/moe_common.py
so the three 2-stage MoE packages share one copy, and rename their private
submodules (_gemm1/_gemm2/_reduction/_common) to drop the underscore. Public
__init__ facades unchanged; generated IR is byte-identical.
