+++ kernels/moe/moe_common.py
@@ -1,9 +1,12 @@
 # SPDX-License-Identifier: MIT
 # Copyright (C) 2024-2026, Advanced Micro Devices, Inc. All rights reserved.
 
-"""Common types shared across MoE FlyDSL kernel modules."""
+"""Common types and helpers shared across MoE FlyDSL kernel modules."""
 
 from enum import Enum
+
+from flydsl.expr import vector
+from flydsl.expr.typing import T
 
 
 class GateMode(str, Enum):
@@ -22,3 +25,39 @@ class GateMode(str, Enum):
     MOCK_GATE_ONLY = "mock_gate_only"
     GATE_ONLY = "gate_only"
     INTERLEAVE = "interleave"
+
+
+# ── Vector bit-reinterpretation helpers ──────────────────────────────────────
+# Thin vector.from_elements + vector.bitcast wrappers that repack packed integer
+# lanes into the vector element types consumed by MFMA. Pure (only their args and
+# the module-level vector / T are referenced), shared by the 2-stage MoE kernels.
+
+
+def i64_to_v4f16(x_i64):
+    """Reinterpret one i64 lane as vector<4xf16>."""
+    v1 = vector.from_elements(T.vec(1, T.i64), [x_i64])
+    return vector.bitcast(T.f16x4, v1)
+
+
+def i64_to_v4i16(x_i64):
+    """Reinterpret one i64 lane as vector<4xi16> (bf16 bit pattern)."""
+    v1 = vector.from_elements(T.vec(1, T.i64), [x_i64])
+    return vector.bitcast(T.i16x4, v1)
+
+
+def i64x2_to_v8f16(lo, hi):
+    """Reinterpret two i64 lanes as vector<8xf16>."""
+    v2 = vector.from_elements(T.i64x2, [lo, hi])
+    return vector.bitcast(T.f16x8, v2)
+
+
+def i64x2_to_v8bf16(lo, hi):
+    """Reinterpret two i64 lanes as vector<8xbf16>."""
+    v2 = vector.from_elements(T.i64x2, [lo, hi])
+    return vector.bitcast(T.bf16x8, v2)
+
+
+def i64x4_to_i32x8(x0, x1, x2, x3):
+    """Reinterpret four i64 lanes as vector<8xi32>."""
+    v4 = vector.from_elements(T.vec(4, T.i64), [x0, x1, x2, x3])
+    return vector.bitcast(T.vec(8, T.i32), v4)
